Add files via upload
The “Adaptive Filtered Miner” is an intelligent online
recommendation system that is used to enhance the experience
of using mobile internet. The system ensures that the platform is
capable of generating personalized notifications that can be
delivered to the user. Intelligent techniques are used to deal with
the user profile and thereby to study the user behavior. Artificial
Neural Networks (ANN) are used here for building the classification
model. Labeling of the attributes is necessary to group similar user
behavior together. The labeled information together with the
corresponding input variables form the training data and is used for
establishing the classification model. The testing set is used to test
the performance of the proposed intelligent mobile personalization
model and to predict the content items correctly.